requirements.txt
google-cloud-speech==0.33.0
python-osc==1.6.6
pyaudio